Added 'get_status' command
